delete entity check if dependent file exists

+++ apps/server/src/app/circuits/circuits.controller.ts
@@ -39,10 +39,12 @@ export class CircuitsController {
   async remove(@Param('id') id: string): Promise<void> {
     const circuit = await this.circuitsService.findOne(id);
     if (circuit) {
-      if (circuit.simulatorPath != '') {
+      if (circuit.simulatorPath != '' && fs.existsSync(circuit.simulatorPath)) {
         fs.unlinkSync(circuit.simulatorPath);
       }
-      fs.unlinkSync(circuit.path);
+      if (fs.existsSync(circuit.path)) {
+        fs.unlinkSync(circuit.path);
+      }
     }
     return this.circuitsService.remove(id);
   }
